Move Table component to common"
