+++ assets/js/common/Table/CollapsibleTableRow.jsx
