+++ assets/js/common/Table/EmptyState.jsx
