+++ assets/js/components/ClusterDetails/AscsErsClusterDetails.jsx
@@ -8,7 +8,7 @@ import { RUNNING_STATES } from '@state/lastExecutions';
 import Button from '@common/Button';
 import PageHeader from '@components/PageHeader';
 import BackButton from '@components/BackButton';
-import Table from '@components/Table';
+import Table from '@common/Table';
 import ListView from '@common/ListView';
 import ProviderLabel from '@components/ProviderLabel';
 import DottedPagination from '@components/DottedPagination';
