+++ assets/js/components/ClusterDetails/HanaClusterDetails.jsx
@@ -7,7 +7,7 @@ import PageHeader from '@components/PageHeader';
 import BackButton from '@components/BackButton';
 
 import ListView from '@common/ListView';
-import Table from '@components/Table';
+import Table from '@common/Table';
 import Tooltip from '@components/Tooltip';
 import ClusterNodeLink from '@components/ClusterDetails/ClusterNodeLink';
 import CheckResultsOverview from '@components/CheckResultsOverview';
