+++ assets/js/components/ClustersList.jsx
@@ -8,7 +8,7 @@ import { getClusterTypeLabel } from '@lib/model/clusters';
 import { addTagToCluster, removeTagFromCluster } from '@state/clusters';
 import { getAllSAPInstances } from '@state/selectors/sapSystem';
 
-import Table from '@components/Table';
+import Table from '@common/Table';
 import Tags from '@components/Tags';
 import PageHeader from '@components/PageHeader';
 import ClusterLink from '@components/ClusterLink';
