+++ assets/js/components/DatabasesOverview/DatabasesOverview.test.jsx
@@ -7,7 +7,7 @@ import userEvent from '@testing-library/user-event';
 
 import { databaseFactory } from '@lib/test-utils/factories';
 import { renderWithRouter } from '@lib/test-utils';
-import { filterTable, clearFilter } from '@components/Table/Table.test';
+import { filterTable, clearFilter } from '@common/Table/Table.test';
 
 import DatabasesOverview from './DatabasesOverview';
