+++ assets/js/components/HealthSummary/HomeHealthSummary.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import Table from '@components/Table';
+import Table from '@common/Table';
 import HealthIcon from '@components/Health/HealthIcon';
 import { Link } from 'react-router-dom';
 import PageHeader from '@components/PageHeader';
